+++ lib/c_runner.rb
